all: Initial commit of original PIC18 code and new MPLABX project
In 2008, I made an implementation of basic USB stack for PIC18F4550 as a
learning exercise.  It enumerated and transferred data, but it wasn't
designed to be a library and so the USB logic was all tied up with the
application.

In 2013, I decided that I needed an open source USB stack for Microchip PIC
MCUs to support my open hardware projects, since the Microchip USB stack
does not allow redistribution of the source.

This is the code that worked on the PIC18F4550 in 2008 (usb.c and usb.h)
with an MPLAB X project file for PIC24.  This will not compile, but is the
starting point for the project.