usb: More race condition fixes (BDnSTAT)
So it appears that on the PIC18F46J50, BDnSTAT must be set as a single
operation, the entire byte at once, not using bit operations.  It'd be
nice if stuff like this were in the datasheet.  This actually makes the
code cleaner looking and results in a smaller footprint anyway.