all: Control transfer support from application (IN and OUT)
Add support for control transfers from the application. This is currently
implemented with a static callback and a requirement that the application
call back into the USB stack from this callback to set up the data stage
(whether IN, OUT or not present).  Once the transfer has completed, the
USB stack will call an application-provided dynamic callback to signal
completion or failure.