usb: Make start_control_return() take a const parameter
Make start_control_return() so that const descriptors can be passed in. The
only downside is that it must cast the const away in order to set
ep0_data_stage_buffer to it.  Since ep0_data_stage_buffer is also used for
receiving data (and thus must be written to), it must be non-const.