usb: Refactor endpoint 0 SETUP packet handling, phase 1
Phase 1 is to copy and paste the code into a new function, without fixing
the indenting.  This way this diff looks clean.  If you notice, this diff
is the first part of usb_service() before it gets to the SETUP handling.
The diff moves the first part of usb_servce() from before the SETUP
handling to after, which is equivalent to moving SETUP handling, but
is expressed in less diff lines.