usb: Support large endpoints
For full-speed devices, isochronous endpoints can be up to 1023 bytes in
length, which is longer than can fit into 8-bits of BDnCNT.  This commit
adds support for using the BC8 and BC9 bits of BDnSTAT to get this extra
range.  The original design was to treat BDnSTAT and BDnCNT as a single
16-bit field, and update the count and the flags with a single C
operation.  Unfortunately, this can't be done on the 8-bit PIC for two
reasons:

1. On 8-bit PIC BDnSTAT and BDnCNT are two different fields, BDnSTAT
first.  This means the BC bits and the BDnCNT bits, while consecutive, are
effectively big-endian, and the XC8 compiler is little-endian (even though
the hardware is 8-bit).  Even if the endian order were right, the XC8
compiler doesn't support bitfields greater than 8-bits in length.

2. BDnSTAT needs to be written _after_ BDnCNT. With a 16-bit variable on
an 8-bit MCU, this is impossible to guarantee from C.

The result is to make a macro to set the flags and the count. On 16-bit
PIC, it sets both in a single operation.  On 8-bit PIC it uses two
operations, one for each byte.  For each platform, two implementations of
this macro are introduced, one which uses all ten bits, and another that
just uses the lower 8 bits, the idea being that for the vast majority of
cases, endpoints will be below 256 bytes in length, and single-byte
operations can be done more efficiently.  At some point a macro will
detect long endpoints and turn this functionality on.  It's not needed
right now as the stack doesn't yet support isochronous endpoints.