bootloader: Perform the reset in the transfer callback
Now that the callback has been fixed for transfers with no data stage,
handle the bootloader's reset command in the callback. This way the
transfer completes before the reset occurs, causing the host to see a
successful transfer (and not report an error because the device was reset
before the transfer was complete).