usb: Send zero-length packet only if returning less than requested
The previous implementation sent a ZLP if the transaction length was
divisible by the endpoint's max packet size.  A ZLP should be sent only if
both the transfer length is divisible by the endpoint's max packet size
and the device is returning fewer bytes than were requested in the setup
packet's wLength field.