usb: Split ep0_data_stage_buffer into in/out buffers
Split ep0_data_stage_buffer into:
	ep0_data_stage_in_buffer
	ep0_data_stage_out_buffer

The XC8 compiler has a problem with pointers which can point to both const
(flash memory) and non-const (RAM) data on the PIC16F (which later commits
will add support for).  The solution is to use two separate pointers for
control transfers, one for IN transfers and one for OUT transfers.