usb: Use start_control_return() for all setup requests
Setup requests which return only a few bytes of data (1-2) would set the
BDN directly and did not go through start_control_return().  This is fine
when ping-pong mode is not active on endpoint 0 IN, but will cause
complication when in full ping-pong mode.

Convert the following setup requests to use start_control_return():
	GET_CONFIGURATION
	GET_STATUS (device and endpoint)
	GET_INTERFACE