usb, usb_hal: Add support for full ping-pong mode (mode 2)
The ping-pong of EP 0 IN is the only thing that was missing.

1. Add macros for BDS*() for this mode.
2. Issue a ping-pong reset in usb_init(). This is so the firmware will be
   in sync with the hardware after the bus reset which follows the first
   part of the device descriptor at the beginning of enumeration.
3. Keep track of and use the EP_TX_PPBI flag for endpoint 0 IN.
4. Create copy_to_ep0_in_buf() function to remove some redundancy and use
   it where appropriate.
5. Call usb_send_in_buffer() from start_control_return() instead of
   manipulating the buffer descriptor manually. This localizes the
   accesses of the buffer descriptor.
6. Add SFR #defines in usb_hal.h

This modification makes the system operate in ping-pong mode, but it
doesn't really take full advantage of ping-pong mode to speed up IN
control transfers.  For example, for transfers containing multiple
transactions, it will still wait for the IN transaction to complete before
giving the next block of data to the SIE.  One could imagine an
optimization where the firmware could load both buffers at the beginning
of a Control IN transfer and then replace each as it was sent, keeping two
in the queue.