usb: Handle banked addresses on PIC16
The USB buffer descriptors need linear addresses, but the XC8 compiler
will generate banked (not linear) addresses for the arrays in ep_buffers
if ep_buffers can fit within a single bank.  This isn't a problem when
large endpoints are used, but for small endpoint sizes, the compiler
generates the banked addresses, causing USB operations to fail.

Add a function to generate the linear addresses from the banked ones. This
function is fairly slow, but only needs to be executed at startup.