bootloader: Delay before resetting
For some reason if you don't delay there will sometimes be an error on the
host side when the reset occurs.  On my test system it has to do with
whether I'm plugged into a hub.  In theory the callback shouldn't be
called until the transaction is over, but I'm seeing failures from the
host side nonetheless.  This delay seems to fix the problem and is what
every other bootloader I've seen does.