usb: Only set the returning_short flag if a return is truly short
Short returns for control transfers need to have a zero-length
packet appended if the return is divisible by the endpoint length.

The previous logic would set the returning_short flag if a requested data
item was simply different than the length requested (either longer or
shorter).  This caused a zero-length packet to be sent in cases where it
wasn't appropriate, causing the next transfer to receive the zero-length
packet and thus fail.