Added a configuration for the P-Star 25K50 Micro to each app.
Added config bits for the PIC18F25K50 and PIC18F45K50 to each example
app (except the bootloader).  However, these config bits will not be
needed by a typical user of the P-Star because its config bits are
controlled by the bootloader rather than the application.

Note: the hid_composite app did not work on the P-Star when I tested it.

Signed-off-by: David (Pololu) <dev-david@pololu.com>