usb_msc: Change write interface to be more flexible
Change the write interface to be flexible in its buffer size and in how
success and failure are handled.

1.  Change the interface to not be based on blocks.  Any buffer size can be
used from the application side as long as it's equal to or longer than the
endpoint length and is a factor of the block size.  Many applications will
use a 64-byte block size.

2. The interface for dealing with writes now has two functions:
a. msc_notify_write_data_handled() - Notify that data was handled, or
   pushed to the medium (or buffered) in some meaningful way.
b. msc_notify_write_operation_complete() - Notify that the entire
   operation has completed.

These fit well (but are not tied to) the MMC multi-block write
implementation.