msc_test: Support multi-block MMC write
This takes advantage of multi-block write support added to the MMC
component (in storage/) and support for smaller-than-block buffer sizes in
the Mass Storage component.  Multi-block writing will increase the life of
an MMC card and will increase performance, since the USB peripheral and
the MMC device can be working better.

The previous implementation waited for the USB/MSC stack to transfer an
entire block of data while the MMC code was idle.  This implementation
allows a write to begin after just 64 bytes have been sent from the host
if desired.  The buffer size is now configurable on the application side.