usb: Don't memcpy() zero bytes
The XC8 compiler, in recent versions (at least 1.34, 1.35, and 1.37) has
developed a regression in which it corrupts memory if memcpy() is called
with zero as the number of bytes.  This happens on PIC16F, but not on
PIC18F.

Check for zero, and don't call memcpy() if there are zero bytes.

Incidentally, this also avoids the technically undefined behavior of
passing NULL pointers into memcpy(), which is not valid by the C standard
even if the size is zero.