usb: Change ep0_data_stage_callback() call location
Change where ep0_data_stage_callback() gets called.  There are three cases
for a control transfer:
1. OUT with no data,
2. OUT with data,
3. IN with data.

Previously, for all three cases, the callback was actually called at the
end of the Status stage.  This commit makes the stack call the callback at
the end of the Data stage in case #2.

The benefit is that now the application can optionally reject the data
stage (by returning -1 from the callback).  This will cause a Stall status
to be sent to the host, notifying the the host of the failure.