bootloader: Add skip regions
Skip regions are regions of flash memory, which may exist in the HEX file,
that the bootloader should simply skip during programing and verification.

The reason for these skip regions is that it is impossible with the XC32
linker to completely eliminate the debug executive, even when not using the
debugger.  The XC32 linker will place this debug executive in the same
hard-coded spot, regardless of what is in the linker script.  It's a real
bummer and a hack, but it's necessary, at least for now.