cdc_acm: Handle getting and setting of line coding
Handling of SET_LINE_CODING and GET_LINE_CODING are required, and Windows
terminals will fail if those requests fail.

This used to not fail on Windows, but that was by accident as there was no
way to reject (stall) the data stage of control transfers.  The behavior
which triggered this failure was likely caused in d1b95bc5c6dadec and then
c685ebbec1cf267.